Add files via upload
These are the different datasets at different stages. This first stage is the pdf containing a list of preservation cases in Denmark - "Fredningsliste.pdf". I then converted this to a word document in order to make changes - "Fredningsliste.docx". I then converted this to en excel document - "Fredningsliste_raw.xlsx." I then tidied the excel document and saved it as "fredningsliste_TIDY.xlsm." In order to upload the file to Studio I saved the final version as a csv file - "fredningsliste_tidy.csv."